+++ database/migrations/2025_08_03_114042_add_role_to_users_table.php
@@ -8,15 +8,18 @@ return new class extends Migration
 {
     public function up(): void
     {
-        Schema::table('users', function (Blueprint $table) {
-            $table->string('role')->after('email')->default('mahasiswa');
-        });
+        // Hanya menghapus kolom 'role'
+        if (Schema::hasColumn('users', 'role')) {
+            Schema::table('users', function (Blueprint $table) {
+                $table->dropColumn('role');
+            });
+        }
     }
 
     public function down(): void
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->dropColumn('role');
+            $table->string('role')->after('email')->default('mahasiswa');
         });
     }
 };
